requirements.txt
requests>=2.25.0
dnspython>=2.1.0
python-whois>=0.7.3
beautifulsoup4>=4.9.0
lxml>=4.6.0
cryptography>=3.4.0
pyOpenSSL>=20.0.0
urllib3>=1.26.0
colorama>=0.4.0
tabulate>=0.8.0